+++ api/v1alpha1/userconfig/service/mysql/mysql.go
@@ -20,6 +20,10 @@ type Migration struct {
 	// Database name for bootstrapping the initial connection
 	Dbname *string `groups:"create,update" json:"dbname,omitempty"`
 
+	// +kubebuilder:validation:Enum="mydumper";"mysqldump"
+	// Experimental! Tool to use for database dump and restore during migration. Default: mysqldump
+	DumpTool *string `groups:"create,update" json:"dump_tool,omitempty"`
+
 	// +kubebuilder:validation:MaxLength=255
 	// Hostname or IP address of the server where to migrate data from
 	Host string `groups:"create,update" json:"host"`