+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - **Upgraded controller-runtime to v0.16.6**: This version supports Kubernetes v0.28 and requires Go 1.20+. See [compatibility matrix](https://github.com/kubernetes-sigs/controller-runtime?tab=readme-ov-file#compatibility) for details.
 - Change `MySQL` field `userConfig.binlog_retention_period`: maximum ~~`604800`~~ → `9007199254740991`
+- Add `MySQL` field `userConfig.migration.dump_tool`, type `string`: Experimental! Tool to use for database
+  dump and restore during migration. Default: mysqldump
 
 ## v0.34.0 - 2025-10-21